changed classes

+++ src/main/java/vat/models/Shape.java
@@ -1,5 +1,6 @@
 package vat.models;
 
 public interface Shape {
-    int calculateContent();
+    double calculateContent();
+
 }
